allow configuration of server at comptime or runtime
This is supposed to be a unique feature of this server, made possible by Zig.
Though, we do still rely on release optimization mode to really optimize
out branch checks.
If we only offered the option of configuration at runtime (like every
other Minecraft server basically) it would not be possible to optimize
out the checks.