simplify Keep Alive ID
Just use 0. There's no point in incrementing it. We don't need every new
Keep Alive ID to be unique.
We do want to probably use milliseconds as the ID in the future though
when we want to properly update the player list showing the ping/latency in
bars to other players (openable with Tab).
We might make this configurable.